+++ packages/core/src/main/Account.ts
@@ -49,7 +49,7 @@ import {AccountService} from './account/';
 import {LinkPreviewService} from './linkPreview';
 import type {CoreCrypto} from '@otak/core-crypto';
 import {WEBSOCKET_STATE} from '@wireapp/api-client/src/tcp/ReconnectingWebsocket';
-import {createCustomEncryptedStore, createEncryptedStore} from './util/encryptedStore';
+import {createCustomEncryptedStore, createEncryptedStore, deleteEncryptedStore} from './util/encryptedStore';
 import {Encoder} from 'bazinga64';
 import {MLSService} from './mls';
 import type {MLSConfig} from './mls/types';
@@ -403,27 +403,42 @@ export class Account<T = any> extends EventEmitter {
     mlsConfig: MLSConfig,
     entropyData?: Uint8Array,
   ) {
+    if (!this.service) {
+      throw new Error('Services are not set.');
+    }
     const coreCryptoKeyId = 'corecrypto-key';
     const {CoreCrypto} = await import('@otak/core-crypto');
-    const dbName = `secrets-${this.generateDbName(context)}`;
+    const dbName = this.generateSecretsDbName(context);
 
     const secretStore = mlsConfig.secretsCrypto
       ? await createCustomEncryptedStore(dbName, mlsConfig.secretsCrypto)
       : await createEncryptedStore(dbName);
 
     let key = await secretStore.getsecretValue(coreCryptoKeyId);
+    let isNewMLSDevice = false;
     if (!key) {
       key = window.crypto.getRandomValues(new Uint8Array(16));
       await secretStore.saveSecretValue(coreCryptoKeyId, key);
+      // Keeping track that this device is a new MLS device (but can be an old proteus device)
+      isNewMLSDevice = true;
     }
+
     const {userId, domain} = this.apiClient.context!;
-    return CoreCrypto.init({
+    const mlsClient = await CoreCrypto.init({
       databaseName: `corecrypto-${this.generateDbName(context)}`,
       key: Encoder.toBase64(key).asString,
       clientId: `${userId}:${client.id}@${domain}`,
       wasmFilePath: mlsConfig.coreCrypoWasmFilePath,
       entropySeed: entropyData,
     });
+
+    if (isNewMLSDevice) {
+      // If the device is new, we need to upload keypackages and public key to the backend
+      await this.service.client.uploadMLSPublicKeys(await mlsClient.clientPublicKey(), client.id);
+      await this.service.client.uploadMLSKeyPackages(await mlsClient.clientKeypackages(this.nbPrekeys), client.id);
+    }
+
+    return mlsClient;
   }
 
   private async registerClient(
@@ -443,11 +458,6 @@ export class Account<T = any> extends EventEmitter {
         this.mlsConfig,
         entropyData,
       );
-      await this.service.client.uploadMLSPublicKeys(await this.coreCryptoClient.clientPublicKey(), registeredClient.id);
-      await this.service.client.uploadMLSKeyPackages(
-        await this.coreCryptoClient.clientKeypackages(this.nbPrekeys),
-        registeredClient.id,
-      );
     }
     this.apiClient.context!.clientId = registeredClient.id;
     this.logger.info('Client is created');
@@ -464,7 +474,15 @@ export class Account<T = any> extends EventEmitter {
     delete this.service;
   }
 
-  public async logout(): Promise<void> {
+  /**
+   * Will logout the current user
+   * @param clearData if set to `true` will completely wipe any database that was created by the Account
+   */
+  public async logout(clearData: boolean = false): Promise<void> {
+    if (clearData && this.coreCryptoClient) {
+      await this.coreCryptoClient.wipe();
+      await deleteEncryptedStore(this.generateSecretsDbName(this.apiClient.context!));
+    }
     await this.apiClient.logout();
     this.resetContext();
   }
@@ -586,6 +604,10 @@ export class Account<T = any> extends EventEmitter {
   private generateDbName(context: Context) {
     const clientType = context.clientType === ClientType.NONE ? '' : `@${context.clientType}`;
     return `wire@${this.apiClient.config.urls.name}@${context.userId}${clientType}`;
+  }
+
+  private generateSecretsDbName(context: Context) {
+    return `secrets-${this.generateDbName(context)}`;
   }
 
   private async initEngine(context: Context): Promise<CRUDEngine> {