+++ packages/core/src/main/conversation/ConversationService/ConversationService.ts
@@ -17,6 +17,7 @@
  *
  */
 
+import {ExternalProposalType} from '@otak/core-crypto';
 import type {APIClient} from '@wireapp/api-client';
 import {
   MessageSendingStatus,
@@ -105,6 +106,7 @@ import {
 } from './ConversationService.types';
 import {Decoder} from 'bazinga64';
 import {mapQualifiedUserClientIdsToFullyQualifiedClientIds} from '../../util/mapQualifiedUserClientIdsToFullyQualifiedClientIds';
+import {optionalToUint8Array} from '../../mls';
 
 export class ConversationService {
   public readonly messageTimer: MessageTimer;
@@ -1272,5 +1274,28 @@ export class ConversationService {
       events: messageResponse?.events || [],
       conversation,
     };
+  }
+
+  /**
+   * Will send an external proposal for the current device to join a specific conversation.
+   * In order for the external proposal to be sent correctly, the underlying mls conversation needs to be in a non-established state
+   * @param conversationGroupId The conversation to join
+   * @param epoch The current epoch of the local conversation
+   */
+  public async sendExternalJoinProposal(conversationGroupId: string, epoch: number) {
+    const groupIdDecodedFromBase64 = Decoder.fromBase64(conversationGroupId!).asBytes;
+    const externalProposal = await this.mlsService.newExternalProposal(ExternalProposalType.Add, {
+      epoch,
+      conversationId: groupIdDecodedFromBase64,
+    });
+    await this.apiClient.api.conversation.postMlsMessage(
+      //@todo: it's temporary - we wait for core-crypto fix to return the actual Uint8Array instead of regular array
+      optionalToUint8Array(externalProposal),
+    );
+  }
+
+  public async isMLSConversationEstablished(conversationGroupId: string) {
+    const groupIdDecodedFromBase64 = Decoder.fromBase64(conversationGroupId!).asBytes;
+    return this.mlsService.conversationExists(groupIdDecodedFromBase64);
   }
 }