+++ packages/core/src/main/mls/MLSService/MLSService.ts
@@ -18,12 +18,19 @@
  */
 
 import {
+  AddProposalArgs,
   CommitBundle,
   ConversationConfiguration,
   ConversationId,
   CoreCrypto,
   DecryptedMessage,
+  ExternalProposalArgs,
+  ExternalProposalType,
+  ExternalRemoveProposalArgs,
   Invitee,
+  ProposalArgs,
+  ProposalType,
+  RemoveProposalArgs,
 } from '@otak/core-crypto/platforms/web/corecrypto';
 import {APIClient} from '@wireapp/api-client';
 import {QualifiedUsers} from '../../conversation';
@@ -115,6 +122,19 @@ export class MLSService {
     }, []);
 
     return coreCryptoKeyPackagesPayload;
+  }
+
+  public async newProposal(
+    proposalType: ProposalType,
+    args: ProposalArgs | AddProposalArgs | RemoveProposalArgs,
+  ): Promise<Uint8Array> {
+    return this.getCoreCryptoClient().newProposal(proposalType, args);
+  }
+  public async newExternalProposal(
+    externalProposalType: ExternalProposalType,
+    args: ExternalProposalArgs | ExternalRemoveProposalArgs,
+  ): Promise<Uint8Array> {
+    return this.getCoreCryptoClient().newExternalProposal(externalProposalType, args);
   }
 
   public async processWelcomeMessage(welcomeMessage: Uint8Array): Promise<ConversationId> {